python: Respect --legacy_external_runfiles for runfiles.
This was found by way of the upb downstream tests, which currently rely on the
legacy external runfiles format for runfiles (this is when files that come from
other repositories are put in `$runfilesRoot/$mainRepo/external/$otherRepo`).

What happens is a regular `runfiles.merge(other)` call will, under the hood,
create a new Runfiles.Builder object with legacyExternalRunfiles=false, thus
losing the value from the flag. To fix, we just have to create a builder with
the value from the flag so that the created Runfiles object keeps it.

Workaround for https://github.com/bazelbuild/bazel/issues/17415

PiperOrigin-RevId: 507503578
Change-Id: I961f50f73ae1ee9980da92eff1b333cbc3b82104