Add feature definition for `static_link_cpp_runtimes` in unix cc toolchain config
`static_link_cpp_runtimes` is one of the well-known features - which is required if you want to link in non-default toolchain libs. However the default unix toolchain does not define it at all.

This PR adds a simple definition of this feature. The added feature is disabled by default to avoid negative impacts, and can be enabled on demand.

Closes #17391.

PiperOrigin-RevId: 507677235
Change-Id: If149a50c1ab41dad258f706bc83df0fc09f8e6e7