Use a concurrent map in ClassToInstanceMapSkyKeyComputeState
Partial reevaluation may mutate this map from different threads. When
the SkyKeyComputeState cache is cleared (e.g., because of memory
pressure), the evaluation of a SkyFunction which opts into partial
reevaluation may check its mailbox at the same time that one of its
children completes and tries to signal that mailbox. The mailbox was
threadsafe, but previously, ClassToInstanceMapSkyKeyComputeState was not.
The two threads could race to construct a new mailbox, which could result
in the checked mailbox constructed by the SkyFunction evaluation thread
getting overwritten by an unchecked mailbox constructed by the signalling
child thread. When that happens,
GenQueryDirectPackageProviderFactory.Function's expectations about its
cached mailbox and its ScopeTraversal states could be violated -- the
next restart could observe a FRESHLY_INITIALIZED mailbox and a non-empty
ScopeTraversal, which should have been impossible.

MutableClassToInstanceMap could not be retained in
ClassToInstanceMapSkyKeyComputeState, because it is implemented using
ForwardingMap, which doesn't provide atomic computeIfAbsent semantics
even if provided with a ConcurrentHashMap delegate.

PiperOrigin-RevId: 507842445
Change-Id: I1ccbeefde751a844dc27ba9ba1a9cf02a6546702