Automated rollback of commit 6224c8754bcfd9f247cd13806052cabf58e2a60b.
*** Reason for rollback ***

No longer needed, change is not good for walltime under prioritization.

This change was originally needed to ensure prioritization information
could propagate, as it had been held within SkyKeys. This caused a ~10MB
heap-after-gc regression. The new approach propagates priority information
through DirtyBuildingState which avoids this memory regression because the
state is cleaned up after evaluation.

With prioritization, there are 2 queues, a priority queue and a FIFO. Going
through the priority queue is much more expensive than the FIFO. So it's better
to not mark too many keys CPUHeavy.

Additionally, requiring the keys to be transitively connected would be a very
difficult property to maintain over time.

*** Original change description ***

Mark SkyKeys having CPUHeavy dependencies CPUHeavy.

* This may improve walltime in high CPU settings.
* This is a precursor to SkyKey prioritization, which requires it.

PiperOrigin-RevId: 509253113
Change-Id: I55479c3196c96c5e9319da7e052cb153cd24229f