python: Enable Starlark implementation of Python rules
This enables the Starlark implementation of the Python rules. While these are
largely drop-in replacements, there are a handful of edge cases that
are known to be incompatible; see https://github.com/bazelbuild/bazel/issues/15897
for the list of cases.

If you believe the Starlark implementation to be the cause of a problem, please
post to https://github.com/bazelbuild/bazel/issues/15897.

Work towards #15897

PiperOrigin-RevId: 509295132
Change-Id: Icbe5beb32e0700a915ac40fe4dbcacf102382974