Split StarlarkList into interface and mutable implementation of it
Replace sameElems with Sequence.sameElems to reduce code duplication.

This is a preparation for other implementations of StarlarkList. In particular, immutable implementation and singleton immutable implementation.

The follow up optimisation of lists implementation is essential for optimisation of Starlark providers.

The performance (cpu and memory) is the same before and after.

PiperOrigin-RevId: 509459853
Change-Id: I0b78344877a24c2be680accb9c1294a6f03ce9e2