Prevent arbitrary modifications to the backing `InMemoryGraph` map by removing the `getAllValuesMutable()` method in favor of more specific methods that don't expose a mutable map.
The primary use case for exposing the graph as a `ConcurrentHashMap` is for the parallel `forEach` method. However, that is now implemented directly on `InMemoryGraph` as `parallelForEach`.

The motivation for this change is for upcoming work on using `InMemoryGraph` as a `SkyKeyPool`, in which the `remove()` method will need to do some additional work to ensure weak interning of keys when they are removed from the pool. This change ensures that the only way to remove keys from the graph is to go through `InMemoryGraph#remove`.

PiperOrigin-RevId: 509605561
Change-Id: I7ec69aeb58c26fa3e17f757fc357514f58480c1e