documentation: clarify combined cache
In Bazel, a Combined Cache is in effect when both Local Disk Cache (--disk_cache=) and Remote Cache (--remote_cache=) are set. This concept, however, only exists in Bazel source code and not defined anywhere in Bazel's documentation.

Let's clarify Combined Cache concept in the doc to explain to users what it is.

Closes #17357.

PiperOrigin-RevId: 509738073
Change-Id: I96c23e5eea086b73a4f0a05c6be7a1e985401369