Allow the timeout value for execute calls on a mac to be set via an environment variable
In certain setups, these calls to xcrun during bazel setup have been reported to
sometimes take more than 2 minutes (see
https://github.com/bazelbuild/bazel/issues/17437).  We have already bumped this
timeout multiple times, which is currently 60 for some calls and 120 for others.
Standardize the default timeout to 120, and allow the timeout to be override via
BAZEL_OSX_EXECUTE_TIMEOUT, to allow individual enviroments to increase that even
more if needed.

PiperOrigin-RevId: 510200188
Change-Id: I664eb7979c4fd2b46ccc87d073f319c1e6041d77