When freezing an `AttributeContainer`, do not store non-explicit values that match the attribute's default value.
This saves a significant amount of space. Consumers of a frozen `AttributeContainer` must retrieve defaults from the `RuleClass` - this was already the case for regular defaults and is now also done for `ComputedDefault` values as well.

Overhaul `AttributeContainerTest` to use `TestParameterInjector` to run tests with both small/large container sizes and mutable/frozen.

PiperOrigin-RevId: 510248106
Change-Id: I4fd66a0afc576038021ffa1aa52c4fafb0571f67