Remove `preloaded_deps` attribute from cc_shared_library
This attribute came out of the design doc for cc_shared_library for a use case
that hasn't materialized. Adding this attribute later would be a compatible
change. If/when a use case does materialize we can easily add it back and fix
any problems with it then.

From a search across many different repositories, it looks like no one has ever
used the attribute.

RELNOTES:none
PiperOrigin-RevId: 510419035
Change-Id: I569d631ad78a78312e348913325438aa7f0256dc