Remove cc_shared_library_permissions
This mechanism came out of the design doc for cc_shared_library but people aren't using it. The idea was to prevent a cc_library target to be exported by cc_shared_libraries that aren't authorized. In reality though this cannot work on two accounts:
1. The mechanism would only prevent a cc_library from being "claimed to be exported" by the cc_shared_library, however a custom Starlark rule would easily bypass this.
2. The rule cannot control the version script the user passes to the linker so the symbols could in practice still be exported.

In general, this mechanism provides a false sense of security, no one is using it and currently the only thing it does is to slightly complicate the codebase and the documentation. In any case adding this functionality later if needed would be a compatible change.

RELNOTES:none
PiperOrigin-RevId: 510425434
Change-Id: Icf85050b57b1d8dae0d32614dc5951d982b3b3d0