Remove stamp attribute from deploy jar rule
Since this rule may be invoked by users directly for java_test, we need to ensure we use the same `stamp` value for both the binary and the deploy jar.

Unfortunately, the bazel implementation uses `java_common.get_build_info()` which looks up stamping on the rulecontext attribute. This means we need to perform this lookup during analysis of the java_binary/java_test.

So we do that and pass the resolved build info files in java_binary/java_test to the deploy jar rule via the internal provider

PiperOrigin-RevId: 510432399
Change-Id: I035858c7ee865cde851b4c67821d69bcbbb7d164