Set right default value for the cc_shared_library flag
Without that flag being True half the motivation for the existence of
cc_shared_library disappears since the rule won't perform any checks for
libraries that are linked more than once. The default must therefore be True.

The flag can be kept around for projects that may want to introduce the
rule into their builds first then solve the duplicated linking issues as a next
step. But usage of the flag is not necessary at all even for existing builds.

The flag was in fact introduced with True as default value but when it was
moved from being written as Starlark configuration to being part of CppOptions
it was mistakenly changed to False.

This change also deletes the obsolete experimental_enable_target_export_check flag.

PiperOrigin-RevId: 511126574
Change-Id: I18f4a63735febea20f1ca1b74d74acdece71740a