Avoid storing a rule's name in `AttributeContainer`.
The rule's name is already accessible from the `Rule#getName` method which looks at the rule's label, so it is not necessary to spend additional storage on the name. To service requests for the name attribute, `Rule` special-cases for `name`, similar to how it already special-cased for `generator_function` and `generator_location`.

For consistency, `RuleClass` now checks that the `name` attribute is always at index 0. This is already the case in practice, since rule classes created through `RuleClass.Builder` add `name` first. The only place that needed updating was `RuleClassTest`, which invokes the `@VisibleForTesting` constructor directly.

PiperOrigin-RevId: 511511515
Change-Id: I3f34bea0e8bdf8bfece767bd5df3933c94a3464d