Optimise Depsets in Starlark providers
Add unit tests for provider optimisation

On schemaful providers store predicted element classes of a Depset next to the schema and encode the actual Depset with a NestedSet.

The prediction is based on the first non-empty Depset in a provider. The optimisation either works or if it doesn't it causes no harm.

PiperOrigin-RevId: 511732456
Change-Id: I1ce34d9d117645429759d49a7fb461becdb71aa7