Add MS_BIND|MS_REC to the mount of `/proc`
Despite the usual meanings of these flags, this isn't actually a bind
mount for `proc`, which doesn't appear to be documented anywhere.

Otherwise it fails if there are any other mounts under the host `/proc`.
I think this is due to the behavior documented on the `mount(2)`
manpage for bind mounting:

> EINVAL In an unprivileged mount namespace (i.e., a mount namespace
> owned by a user namespace that was created by an unprivileged user), a
> bind  mount  operation  (MS_BIND) was attempted without specifying
> (MS_REC), which would have revealed the filesystem tree underneath
> one of the submounts of the directory being bound.

I ended up in this scenario with docker and `--gpus all` (with some
additional flags to make it work even without the `--gpus all`).