Correctly set default subprocess factory when loading class `SubprocessBuilder`. (#17578)
Currently, there are two subprocess factories: `JavaSubprocessFactory` and `WindowsSubprocessFactory`. The default factory is set to `JavaSubprocessFactory` when loading class `SubprocessBuidler`.

We call `setDefaultSubprocessFactory` to set the default factory, when creating BlazeRuntime, by checking the OS, which set the default factory to WindowsSubprocessFactory on Windows. However, for integration tests written in Java, `setDefaultSubprocessFactory` is not called. So we use `JavaSubprocessFactory` even on Windows for these tests which doesn't correctly terminate process tree.

This CL fixes that by calling `setDefaultSubprocessFactory` when loading class `SubprocessBuilder`.

PiperOrigin-RevId: 511810605
Change-Id: I55ca32caa0f81160e027780d671bf46a6bd6c266