Add exec_kind to RunnerCount in BEP
The execution kind of runners is now stored
in the RunnerCount message of the Build Event
Protocol. For runners where execution kind is
not applicable, e.g. "total", no value is set.
The exec_kind field has been set as a string
field rather than using an enum to avoid
implicit dependencies between the Java code and
the protobuf.

Change-Id: Ic3aa8330f2f5ac68a7fd6ec2b242fba43265ed71