Correctly handle templated tree artifacts in the prefetcher.
Due to the existence of templated tree artifacts (tree artifacts where each file is produced by a separate action) and nested artifacts (artifacts whose output path is a descendant of another artifact's), the prefetcher must not assume every tree artifact is fetched as a whole.

In addition, concurrent actions calling the prefetcher might write to the same directory in the output tree, so they must synchronize when making the directory temporarily writable.

Fixes #16333.

PiperOrigin-RevId: 513205572
Change-Id: I827c4643643f63c9425e63bdf9177805c5f0f409