Fix `android_binary` native libraries linking check
Prior to this change, the `containsCodeToLink` check behaved differently for libraries with object files unset and libraries with an empty list of object files. The former case was treated as an opaque library and always linked in. In the latter case, the library was never linked (this case only occurs for `cc_library` targets without sources which generate an empty archive).

With this change, we treat both cases as the same, and only link in libraries if their archive is a source artifact i.e. a pre-compiled archive that has been checked-in to the depot.

This change is not expected to alter any observable behavior, and only removes the reliance on `null`-ness of fields in `LibraryToLink` unblocking Starlarkification of `android_binary`

PiperOrigin-RevId: 513501308
Change-Id: I377fa3f60b22d15bdfc7956f181529300e787384