Update `NestMembers` attributes to remove pruned classes
Following the example of `InnerClasses`. This was being worked around by pinning
the test target to compile as Java 8, which disabled the use of the `NestMembers`
attribute (unknown commit).

This change makes ijar handle the attribute correctly, and causes references to
local classes to be pruned when compiling as Java 11.

PiperOrigin-RevId: 513560040
Change-Id: I2ba6a33501f7d6f4e022bf4b523ae9dd238cbea8