Switch `RemoteFileArtifactValue` subclassing to optimize for memory cost.
Adding `long expireAtEpochMilli` added 8 bytes to every `RemoteFileArtifactValue`, even ones without an expiration time. In contrast, `PathFragment materializationExecPath` is just a pointer so only costs 4 bytes, and since without it the cost is rounded up to the next multiple of 8 anyway (28 -> 32), it is free to  add this field.

Conveniently, since the expiration doesn't factor into `hashCode`/`equals`, it's actually simpler to arrange the classes this way.

PiperOrigin-RevId: 514388630
Change-Id: I50a7e4e03b266f1a2dfad2fa38005c0c9a0cfa10