Set runners for critical path where no spawn was actually executed.
Not entirely sure that this does the right thing in all circumstances. Is it possible that an actual spawn runner's update could happen after the call to `finishActionExecution`? What happens around action rewinding? Could we do something about the fileset and symlinking actions that only come through `actionComplete`? That even happens for all actions, so it should not be considered for "non-runner name". Should the runners that are not "real" runners be indicated in some way, in brackets or something?

PiperOrigin-RevId: 514405743
Change-Id: I0323f930a0e3ee475fd07d39c375edcfc542fc96