python: remove unused Java code
With the deletion of the Java rule implementation, large amounts of code are
now unused. Entire removal of the classes isn't yet possible, but most of
their code can be removed.

This also makes it easier to transition to the Starlark providers, since less
Java code is referencing the to-be-deleted Java providers

Work towards #15897

PiperOrigin-RevId: 514566571
Change-Id: I28ef124369fc678945873bc21919830cb73eafc8