Change the internal representation of `GroupedDeps` to save space.
Compressed deps are stored in a flat array. Groups are delimited with size markers. This has much less storage cost than using an array of lists.

To keep compression and decompression fast, I elected to change the uncompressed representation as well. All operations maintain the same time complexity as the nested list representation.

PiperOrigin-RevId: 514619932
Change-Id: I0ab0acbbcef672638b22ab4a93a41401f75cad02