[Skymeld] Drop the intermediate states of IncrementalPackageRoots upon analysis completion.
By holding on to these Packages, we prevent them from being garbage collected,
which makes the uncollectable portion of the heap larger. This leads to more
GCs attempts, since there's less available heap for the rest of the work.

The multiple GCs result in objects being promoted to Old Gen faster, and
eventually costlier GC overall.

PiperOrigin-RevId: 514710229
Change-Id: I1b3ddb6dfb587b778f078837f5faf7a0fd87afe3