Fix cc_shared_library to take into account indirect top level deps
The cc_shared_library implementation wraps its top level direct dependencies in
whole archive because otherwise since there aren't any symbols in the shared
library depending on them they would be dropped by the linker.

There are instances where the target placed in the direct dep of a
cc_shared_library doesn't have a linker_input or if it has it, the linker_input
may not have any code to be linked statically into the shared library.

When that happens, the actual code to be linked is depended on indirectly, e.g.
cc_shared_library -> cc_proto_library -> proto_library. Here it is the code
generated by the aspect applied on proto_library that should be linked
statically. Since this dependency was indirect, that code wasn't
whole-archived. Before this fix, we only tried to whole-archive
cc_proto_library but the cc_proto_library target itself did not produce any
linker_inputs, for this reason the code from the proto_library was dropped by
the linker.

This CL also adds more test coverage for the dynamic_only_roots failure
triggered explicitly by the cc_shared_library implementation.

RELNOTES:none
PiperOrigin-RevId: 515017649
Change-Id: I022e1cb4f70fa4584893d3c34172236fdf2e5f73