Have getFdoBuildStamp return "SAFDO" when a .afdo profile is supplied along with xbinary_fdo.
This change enables the fdo_type build stamp to be set to SAFDO in order to distinguish and track the prevalence of SAFDO in the fleet. As an alternative to this implementation, I also considered defining an SAFDO build feature and using a new file extention to identify SAFDO file. This CL is here unknown commit. I decided this implementation is preferable since from a blaze perspective, SAFDO is really no different that AFDO and adding a SAFDO build feature resulted in a huge amount of duplicate boilerplate code. The method used in this CL has the downside of relying on an implementation detail of SAFDO (SAFDO is afdo with xbinary_fdo set) but I think it is overall better due to to how much smaller a change it is but I am open to feedback.

See[]

PiperOrigin-RevId: 515109461
Change-Id: Idc515b058eaf0659556eb381a44fe3618e8eb435