Force the Bazel server Java runtime to use the root locale
This ensures consistent behavior of string operations even if the
individual operations do not set a locale.

Without this change, Bazel can't operate in e.g. a Turkish locale, where
it fails with error messages such as:

In rule 'test', size 'medium' is not a valid size.

This is because Turkish case mapping rules make it so that a capital
ASCII 'I' lowercases to a non-ASCII variant of 'i'.