Implement MetadataProvider in RemoteActionFileSystem.
This makes it possible (in a followup change) to delete the downloadFile()
prefetcher method in favor of prefetchFiles().

When the action filesystem implements MetadataProvider, Skyframe considers it the authoritative source for input metadata [1]. Therefore, getInput() and getMetadata() need to look in all of the right places (the action input map, the output mapping, the remote output tree, and the local filesystem) and get all of the combinations of source/output and local/remote right.

[1] https://cs.opensource.google/bazel/bazel/+/master:src/main/java/com/google/devtools/build/lib/skyframe/SkyframeActionExecutor.java;l=859;drc=9f8395661b2a74299cf01b1a1cdd04cc4f25dcf0

PiperOrigin-RevId: 515642065
Change-Id: I62caf3aafa4acc23f70cc4511997335e82ebb12a