[Skymeld] Don't propagate every event.
Back in https://github.com/bazelbuild/bazel/commit/f7b8b7273bc65c655908081f41302a41a2cb3ff6, this filter was excluded from Skymeld mostly because of the filter's description. This was a mistake. Even without the boundary between analysis and execution in the traditional meaning, it still makes sense that events from an analysis node should not be propagated to its execution dependencies.

This error resulted in an explosion in the number of SkyValues with Events attached to it. These events are then replayed in the subsequent incremental build.

PiperOrigin-RevId: 516168703
Change-Id: I0a05a1f4a5beee84916ee0f0f41d79c3807304d0