[Skymeld] Drop the syscall cache during analysis finished callback.
This frees up some heap space and as a result GC cost.

There's another CL in the work to introduce a slider to skymeld: a value that tells skyframe to hold up execution until a certain percentage of analysis is done. The higher this value is (that means a reduced degree of analysis-execution overlap), the more important it becomes to clear the syscall cache.

PiperOrigin-RevId: 516194981
Change-Id: Iec332d1c2ce9bcbb138d4cc97a4176432c046826