Inline `AttributeContainer` into `Rule`.
There is a one-to-one relationship between these two classes. Combining them saves memory by cutting out the per-object cost. `Rule` has a net change of +1 field which offsets some of the benefit, but I plan on optimizing `Rule` in the near future.

The only significant change from the mechanics in `AttributeContainer` is that the `BitSet` used in the mutable case is replaced with a `byte[]` to keep field types the same as in the frozen case.

Attribute-related methods in `Rule` are grouped together.

PiperOrigin-RevId: 516398753
Change-Id: I0a2b479e40910ca3dced329e08a4779e35a27492