[Skymeld] Introducing `--experimental_skymeld_analysis_overlap_percentage`.
This flag allows us to influence the degree of overlapping between analysis-execution in blaze.

1. Performance

Our benchmarks showed that allowing analysis and execution to overlap freely does not always mean improve performance:
- Increased resource contention: more inflight evaluations means more states to keep and more contention.
- Prolonged analysis phase: the analysis work in Skymeld mode generally finishes much later into the build. This means that we have to hold on to analysis-specific states for much longer. This negatively affects build performance (smaller available heap space -> more GC, increased risk of OOM).

The original goal of Skymeld is to make use of the available CPU resources observed _towards the end_ of the analysis phase. With `--experimental_skymeld_analysis_overlap_percentage=5`, we would allow blaze to finish (roughly) 95% of the analysis phase work [1] before letting the execution work begin.

2. A simulation of noskymeld mode.

With the flag value set to 0, we can enforce that there's _no_ overlapping between analysis/execution. This would allow us to simulate the noskymeld behavior with skymeld enabled. This can be used to e.g. investigate inefficiencies with Skymeld implementation, and may be a way to get around blockers (e.g. multiple package roots).

[1] "roughly" because it's 95% of the top level target, but individually they may take different amount of time to finish.

PiperOrigin-RevId: 516543887
Change-Id: I51cdb3dca9221c0d7187942ab8705aa27d1f4963